moved project area postproc. to view to run sync for tests
